c/r: move criu code to its own file
Trying to cage the beast that is lxccontainer.c.

Acked-by: Serge E. Hallyn <serge.hallyn@ubuntu.com>
